Year2/DPIIS-lab3_v.1.0.1
> The creatures are divided into types

+++ Year-2/DPIIS-designing-programs-in-intelligent-systems/DPIIS-lab3/WorldDirector.cpp
@@ -8,8 +8,6 @@ WorldDirector::WorldDirector()
 
 World* WorldDirector::create_new_world()
 {
-	//std::cout << "\x1b[31;1mCreating\x1b[0m a new \x1b[32;1mworld\x1b[0m!\n";
-	//std::cout << "Creating a new world\n";
 	this->count_of_created_worlds++;
 	World* world = new World(this->count_of_created_worlds);
 	world->world_id = this->count_of_created_worlds;
@@ -24,7 +22,7 @@ World* WorldDirector::create_new_world(std::string file)
 
 void WorldDirector::delete_world(World* world)
 {
-	for (int world_index = 0; world_index < this->worlds.size(); world_index++) {
+	for (int world_index = 0; world_index < (int)this->worlds.size(); world_index++) {
 		if ((world == this->worlds[world_index]) || (world->world_id == this->worlds[world_index]->world_id))
 		{
 			delete this->worlds[world_index];
@@ -37,7 +35,7 @@ void WorldDirector::delete_world(World* world)
 
 WorldDirector::~WorldDirector()
 {
-	for (int world_index = 0; world_index < this->worlds.size(); world_index++) {
+	for (int world_index = 0; world_index < (int)this->worlds.size(); world_index++) {
 		delete this->worlds[world_index];
 		this->worlds[world_index] = NULL;
 	}
